requirements/base.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.0.7
django-channels==0.7.0
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-extensions==2.2.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-password-reset==2.0
django-vanilla-views==1.0.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
oauthlib==3.0.2
pagarme-python==3.3.0
Pillow==6.1.0
psycopg2==2.8.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.3
xmltodict==0.12.0
